services/search/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
rank-bm25
psycopg2-binary
numpy